agent-demo-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.10.0
httpx==0.25.0
boto3==1.40.56
python-dotenv
